Update:
- Only 1 activity on place: HostActivity.kt
- Fix crash when "save" button is hit and the edit text inputs are empty
- Usage of ConstraintLayout or LinearLayout justified
- Databinding implementation + 2 way databinding
- Menu added on ShoeListFragment

+++ app/src/main/java/com/rosalynbm/shoestoreinventory/models/Shoe.kt
@@ -7,6 +7,6 @@ import androidx.room.PrimaryKey
 data class Shoe(@PrimaryKey
                 var id: Int,
                 var name: String,
-                var size: Double,
+                var size: Int,
                 var company: String,
                 var description: String)
